2026-atai/containers/langflow-vis/requirements.txt
fastapi
uvicorn
requests>=2.31.0
python-dotenv>=1.0.0
